Move to a suggestion package, cleanup run method, support new alpha/project separation
